I added some buttons to the UI to handle moving the ships. I had to change the positioning of the other buttons to make it look a bit better than it did.
Once the buttons were there I added a new route and MovementGameAction controller to handle the movement of the ships on the server. Once this was in place I added some event handlers to the buttons that simply sent a MovementGameAction to the server, which is handled inside the ApplicationController. I haven't added the functionality to actually move the ships yet, but I did make it so the server prints out the direction the user wants to move ships just to make sure it is properly sending the data.