still working on ship movement
managed to get ships moving, but only sometimes. The way I am checking if the move is valid does not work properly since it is detecting itself as a ship that is in it's place and returning false. I am going to write a new function for checking if the move is valid.